+++ inc/modules/event-list/assets/templates/event-list.php
@@ -1,55 +1,51 @@
 <?php
 namespace BrownPaperTickets;
 
-require_once( plugin_dir_path( __FILE__ ).'../lib/bptWordpress.php' );
+use \BrownPaperTickets\BptWordpress as Utilities;
 
-use BrownPaperTickets\BptWordpress;
+$show_prices = get_option( '_bpt_show_prices' );
+$show_dates = get_option( '_bpt_show_dates' );
+$show_full_description = get_option( '_bpt_show_full_description' );
+$show_location_after_description = get_option( '_bpt_show_location_after_description' );
+$shipping_countries    = get_option( '_bpt_shipping_countries' );
+$shipping_methods = get_option( '_bpt_shipping_methods' );
+$currency = get_option( '_bpt_currency' );
+$date_format = esc_html( get_option( '_bpt_date_format' ) );
+$time_format = esc_html( get_option( '_bpt_time_format' ) );
+$show_end_time = get_option( '_bpt_show_end_time' );
+$event_list_style = get_option( '_bpt_event_list_style' );
 
-$_bpt_show_prices = get_option( '_bpt_show_prices' );
-$_bpt_show_dates = get_option( '_bpt_show_dates' );
-$_bpt_show_full_description = get_option( '_bpt_show_full_description' );
-$_bpt_show_location_after_description = get_option( '_bpt_show_location_after_description' );
-$_bpt_shipping_countries    = get_option( '_bpt_shipping_countries' );
-$_bpt_shipping_methods = get_option( '_bpt_shipping_methods' );
-$_bpt_currency = get_option( '_bpt_currency' );
-$_bpt_date_format = esc_html( get_option( '_bpt_date_format' ) );
-$_bpt_time_format = esc_html( get_option( '_bpt_time_format' ) );
-$_bpt_show_end_time = get_option( '_bpt_show_end_time' );
-$_bpt_event_list_style = get_option( '_bpt_event_list_style' );
-
-//$_bpt_event_list_template = get_option( '_bpt_show_event_list_template' );
-
-if ( $_bpt_date_format === 'custom' ) {
-	$_bpt_date_format = esc_html( get_option( '_bpt_custom_date_format' ) );
+if ( $date_format === 'custom' ) {
+	$date_format = esc_html( get_option( '_bpt_custom_date_format' ) );
 }
 
-if ( $_bpt_time_format === 'custom' ) {
-	$_bpt_time_format = esc_html( get_option( '_bpt_custom_time_format' ) );
+if ( $time_format === 'custom' ) {
+	$time_format = esc_html( get_option( '_bpt_custom_time_format' ) );
 }
 
-if ( $_bpt_currency === 'cad' ) {
-	$_bpt_currency = 'CAD$ ';
+if ( $currency === 'cad' ) {
+	$currency = 'CAD$ ';
 }
 
-if ( $_bpt_currency === 'usd' ) {
-	$_bpt_currency = '$';
+if ( $currency === 'usd' ) {
+	$currency = '$';
 }
 
-if ( $_bpt_currency === 'gbp' ) {
-	$_bpt_currency = '£';
+if ( $currency === 'gbp' ) {
+	$currency = '£';
 }
 
-if ( $_bpt_currency === 'eur' ) {
-	$_bpt_currency = '€';
+if ( $currency === 'eur' ) {
+	$currency = '€';
 }
 
-$countries = BptWordpress::get_country_list();
+$countries = Utilities::get_country_list();
 
-if ( $_bpt_event_list_style ) {
-	$use_style = ( isset( $_bpt_event_list_style['use_style'] ) ? true : false );
+if ( $event_list_style ) {
+	$use_style = ( isset( $event_list_style['use_style'] ) ? true : false );
 
 	if ( $use_style ) {
-		$css = '<style type="text/css">' . esc_html( $_bpt_event_list_style['custom_css'] ) . '</style>';
+		$css = '<style type="text/css">' . esc_html( $event_list_style['custom_css'] ) . '</style>';
 	}
 }
 
@@ -69,7 +65,7 @@ if ( isset( $css ) ) {
 <div class="bpt-loading-<?php esc_attr_e( $post->ID );?> hidden">
 	Loading Events
 	<br />
-	<img src="<?php echo esc_url( plugins_url( 'public/assets/img/loading.gif', dirname( __FILE__ ) ) ); ?>">
+	<img src="<?php echo esc_url( plugins_url( 'assets/img/loading.gif', dirname( __DIR__ ) ) ); ?>">
 </div>
 
 <div id="bpt-event-list-<?php esc_attr_e( $post->ID );?>" class="bpt-event-list" data-post-id="<?php esc_attr_e( $post->ID );?>">
@@ -77,14 +73,20 @@ if ( isset( $css ) ) {
 </div>
 <script type="text/html" id="bpt-event-template">
 
+{{ #error }}
+	<div intro="slide" class="bpt-event bpt-default-theme">
+	<h2>Sorry, an error has occured while loading events.</h2>
+	<p>{{ bptError }}</p>
+	</div>
+{{ /error }}
 
-{{ #bptEvents }}
+{{ #events }}
 	{{ ^.error }}
 	<div intro="slide" class="bpt-event bpt-default-theme">
-		<h2 class="bpt-event-title">{{{ unescapeHTML(title) }}}</h2>
+		<h2 id="bpt-event-{{id}}" class="bpt-event-title">{{{ unescapeHTML(title) }}}</h2>
 
 
-		<?php if ( $_bpt_show_location_after_description === 'false' ) { ?>
+		<?php if ( $show_location_after_description === 'false' ) { ?>
 
 			<div class="bpt-event-location">
 				<div class="address1">{{ address1 }}</div>
@@ -103,7 +105,7 @@ if ( isset( $css ) ) {
 			</p>
 		</div>
 
-		<?php if ( $_bpt_show_full_description === 'false' ) { ?>
+		<?php if ( $show_full_description === 'false' ) { ?>
 		<p>
 			<a href="#" class="bpt-show-full-description" on-click="showFullDescription">Show Full Description</a>
 		</p>
@@ -119,7 +121,7 @@ if ( isset( $css ) ) {
 
 		<?php }
 
-	if ( $_bpt_show_location_after_description === 'true' ) { ?>
+	if ( $show_location_after_description === 'true' ) { ?>
 
 			<div class="bpt-event-location">
 				<div class="address1">{{ address1 }}</div>
@@ -131,23 +133,23 @@ if ( isset( $css ) ) {
 
 		<?php }
 
-	if ( $_bpt_show_dates === 'true' ) { ?>
+	if ( $show_dates === 'true' ) { ?>
 			<form data-event-id="{{ id }}" data-event-title="{{ title }}" method ="post" class="add-to-cart" action="https://www.brownpapertickets.com/addtocart.html" target="_blank">
 				<input type="hidden" name="event_id" value="{{ id }}" />
 				<div class="event-dates">
 					<label for="dates-{{ id }}">Select a Date:</label>
 					<select class="bpt-date-select" id="dates-{{ id }}" value="{{ .selectedDate }}">
 					{{ #dates }}
-						<option class="event-date" value="{{ . }}" >
-							{{ formatDate( '<?php esc_attr_e( $_bpt_date_format ); ?>', dateStart ) }}
-							{{ formatTime( '<?php esc_attr_e( $_bpt_time_format ); ?>', timeStart ) }}
-							<?php echo ( $_bpt_show_end_time === 'true' ? 'to {{ formatTime( \'' . $_bpt_time_format . '\', timeEnd ) }}' : '' ); ?>
+						<option class="event-date" value="{{ . }}">
+							{{ formatDate( '<?php esc_attr_e( $date_format ); ?>', dateStart ) }}
+							{{ formatTime( '<?php esc_attr_e( $time_format ); ?>', timeStart ) }}
+							<?php echo ( $show_end_time === 'true' ? 'to {{ formatTime( \'' . $time_format . '\', timeEnd ) }}' : '' ); ?>
 						</option>
 					{{ /dates }}
 					</select>
 				</div>
 				<fieldset>
-				{{ #selectedDate }}
+				{{ #.selectedDate }}
 					<input name="date_id" value="{{ id }}" type="hidden">
 					<table id="price-list-{{ id }}" class="bpt-event-list-prices">
 					<tr>
@@ -157,31 +159,34 @@ if ( isset( $css ) ) {
 					</tr>
 					{{ #prices }}
 					<tr data-price-id="{{ id }}" class="{{ isHidden(hidden) }}" >
-						<td class="bpt-price-name" data-price-name="{{name}}">
+						<td class="bpt-price-name" data-price-name="{{name}}" data-event-title="{{ title }}">
 						{{ name }}
 						{{ #hidden }}
-							<?php echo ( BptWordpress::is_user_an_admin() ? '<br/><a href="#" on-click="unhidePrice" class="bpt-unhide-price" data-price-name="{{ name }}" data-price-id="{{ id }}">(Display Price)</a>' : '' ); ?>
+							<?php echo ( Utilities::is_user_an_admin() ? '<br/><a href="#" on-click="unhidePrice" class="bpt-unhide-price" data-price-name="{{ name }}" data-price-id="{{ id }}">(Display Price)</a>' : '' ); ?>
 						{{ /hidden}}
-						
+
 						{{ ^hidden }}
-							<?php echo ( BptWordpress::is_user_an_admin() ? '<br/><a href="#" on-click="hidePrice" class="bpt-hide-price" data-price-name="{{ name }}" data-price-id="{{ id }}">(Hide Price)</a>' : '' ); ?>
+							<?php echo ( Utilities::is_user_an_admin() ? '<br/><a href="#" on-click="hidePrice" class="bpt-hide-price" data-price-name="{{ name }}" data-price-id="{{ id }}">(Hide Price)</a>' : '' ); ?>
 						{{ /hidden }}
 
 						</td>
-						<td class="bpt-price-value" data-price-value="{{ formatPrice(value, '<?php esc_attr_e( $_bpt_currency ); ?>' ) }}">{{ formatPrice(value, '<?php esc_attr_e( $_bpt_currency ); ?>' ) }}</td>
+						<td class="bpt-price-value" data-price-value="{{ formatPrice(value, '<?php esc_attr_e( $currency ); ?>' ) }}">{{ formatPrice(value, '<?php esc_attr_e( $currency ); ?>' ) }}</td>
 						<td>
 							<select class="bpt-price-qty" name="price_{{ id }}" data-price-id="{{ id }}">
-
-		<?php
-		$shipping_incr = 0;
-
-		while ( $shipping_incr <= 50 ) {
-			echo '<option value="' . $shipping_incr . '">' . $shipping_incr . '</option>';
-			$shipping_incr++;
-		}
-		?>
+								{{{ getQuantityOptions( . ) }}}
+								<option value="0" selected="true">0</option>
 							</select>
+
+							<?php if ( Utilities::is_user_an_admin() ) { ?>
+							<label class="bpt-price-max-quantity" for="bpt-price-max-quantity-{{ id }}">Set Max Quantity</label>
+							<input id="bpt-price-max-quantity-{{ id }}" type="text" value="{{ maxQuantity }}" class="bpt-price-max-quantity" on-change="setPriceMaxQuantity" placeholder="Max Qty">
+							<?php } ?>
 						</td>
+					</tr>
+					{{ / }}
+					{{ ^prices }}
+					<tr>
+						<td>Sorry, no prices available for the selected date.</td>
 					</tr>
 					{{ / }}
 					</table>
@@ -189,7 +194,7 @@ if ( isset( $css ) ) {
 						<label class="bpt-shipping-method" for="shipping_{{ id }}">Delivery Method</label>
 						<select class="bpt-shipping-method" id="shipping_{{ id }}" name="shipping_{{ id }}">
 		<?php
-		foreach ( $_bpt_shipping_methods as $shipping_method ) {
+		foreach ( $shipping_methods as $shipping_method ) {
 			switch ( $shipping_method ) {
 				case 'print_at_home':
 					echo '<option value="5">Print-At-Home (No Additional Fee)</option>';
@@ -233,28 +238,23 @@ if ( isset( $css ) ) {
 					<div class="bpt-add-to-cart">
 						<button class="bpt-submit" type="submit">Add to Cart</button>
 						<span class="bpt-cc-logos">
-							<img src="<?php echo esc_url( plugins_url( 'public/assets/img/visa_icon.png', __DIR__ ) ); ?>" />
-							<img src="<?php echo esc_url( plugins_url( 'public/assets/img/mc_icon.png', __DIR__ ) ); ?>" />
-							<img src="<?php echo esc_url( plugins_url( 'public/assets/img/discover_icon.png', __DIR__ ) ); ?>" />
-							<img src="<?php echo esc_url( plugins_url( 'public/assets/img/amex_icon.png', __DIR__ ) ); ?>" />
+							<img src="<?php echo esc_url( plugins_url( 'img/visa_icon.png', __DIR__ ) ); ?>" />
+							<img src="<?php echo esc_url( plugins_url( 'img/mc_icon.png', __DIR__ ) ); ?>" />
+							<img src="<?php echo esc_url( plugins_url( 'img/discover_icon.png', __DIR__ ) ); ?>" />
+							<img src="<?php echo esc_url( plugins_url( 'img/amex_icon.png', __DIR__ ) ); ?>" />
 						</span>
 					</div>
 				</div>
 			</form>
 		<?php } ?>
 		<div class="bpt-powered-by">
-			<a href="http://www.brownpapertickets.com/event/{{ id }}" target="_blank"><span>View Event on </span><img src="<?php echo esc_url( plugins_url( 'public/assets/img/bpt-footer-logo.png', __DIR__ ) ); ?>" /></a>
+			<a href="http://www.brownpapertickets.com/event/{{ id }}" target="_blank"><span>View Event on </span><img src="<?php echo esc_url( plugins_url( 'img/bpt-footer-logo.png', __DIR__ ) ); ?>" /></a>
 		<div>
 	</div>
 	{{ /.error }}
-{{ /bptEvents }}
+{{ /events }}
 
 
-{{ #bptError }}
-	<div intro="slide" class="bpt-event bpt-default-theme">
-	<h2>Sorry, an error has occured while loading events.</h2>
-	<p>{{ error }}</p>
-{{ /bptError }}
 
 </script>
 <?php